+++ browser/src/robocorp/browser/__init__.py
@@ -11,7 +11,7 @@ from playwright.sync_api import (
 
 from ._types import BrowserEngine, BrowserNotFound, InstallError
 
-__version__ = "2.2.0"
+__version__ = "2.2.1"
 version_info = [int(x) for x in __version__.split(".")]
 
 
@@ -57,6 +57,13 @@ def configure(**kwargs) -> None:
             new automation).
 
         viewport_size: Size to be set for the viewport. Specified as tuple(width, height).
+
+        skip_playwright_stop:
+            Can be used to skip the playwright stop. Not recommended in general,
+            only meant to be used to diagnose and workaround specific issues on
+            the playwright stop coupled with an early os._exit shutdown in
+            `robocorp-tasks`. Can cause a process leak and even a shutdown
+            deadlock if used alone.
 
     Note:
         See also: `robocorp.browser.configure_context` to change other