+++ windows/src/robocorp/windows/_inspect.py
@@ -189,7 +189,10 @@ class _TkHandler:
 
     @property
     def _default_root(self):
-        return self._roots[0]
+        try:
+            return self._roots[0]
+        except IndexError:
+            return None
 
     def loop(self, on_loop_poll_callback: Optional[Callable[[], Any]]):
         assert self._current_thread == threading.current_thread()
@@ -201,11 +204,17 @@ class _TkHandler:
             def check_action():
                 # Keep calling itself
                 on_loop_poll_callback()
-                self._default_root.after(poll_5_times_per_second, check_action)
+                default_root = self._default_root
+                if default_root is not None:
+                    default_root.after(poll_5_times_per_second, check_action)
 
-            self._default_root.after(poll_5_times_per_second, check_action)
+            default_root = self._default_root
+            if default_root is not None:
+                default_root.after(poll_5_times_per_second, check_action)
 
-        self._default_root.mainloop()
+        default_root = self._default_root
+        if default_root is not None:
+            default_root.mainloop()
 
     def quit(self):
         assert self._current_thread == threading.current_thread()